requirement.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.1
demjson==2.2.4
Flask==1.1.1
future==0.18.2
idna==2.8
itchat==1.2.32
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
py==1.8.1
pycparser==2.19
pyOpenSSL==19.1.0
pypng==0.0.20
PyQRCode==1.2.1
requests==2.22.0
retry==0.9.2
selenium==2.48.0
six==1.14.0
urllib3==1.25.7
virtualenv==16.7.9
Werkzeug==0.16.1
wxpy==0.3.9.8
